Consolidate conftest approach
