+++ vizro-core/tests/unit/vizro/charts/test_charts_utils.py
@@ -1,6 +1,18 @@
 from copy import deepcopy
 
+import plotly.graph_objects as go
 import pytest
+
+from vizro.models.types import capture
+
+
+@pytest.fixture
+def minimal_capture_chart(gapminder):
+    @capture("graph")
+    def _minimal_capture_chart(data_frame):
+        return go.Figure()
+
+    return _minimal_capture_chart(gapminder)
 
 
 @pytest.mark.parametrize("chart_type", ["standard_px_chart", "minimal_capture_chart"])
