+++ vizro-core/tests/unit/vizro/conftest.py
@@ -1,10 +1,8 @@
 """Fixtures to be shared across several tests."""
-import plotly.graph_objects as go
 import pytest
 
 import vizro.models as vm
 import vizro.plotly.express as px
-from vizro.models.types import capture
 
 
 @pytest.fixture
@@ -23,20 +21,6 @@ def standard_px_chart(gapminder):
         hover_name="country",
         size_max=60,
     )
-
-
-@pytest.fixture
-def standard_go_chart(gapminder):
-    return go.Figure(data=go.Scatter(x=gapminder["gdpPercap"], y=gapminder["lifeExp"], mode="markers"))
-
-
-@pytest.fixture
-def minimal_capture_chart(gapminder):
-    @capture("graph")
-    def _minimal_capture_chart(data_frame):
-        return go.Figure()
-
-    return _minimal_capture_chart(gapminder)
 
 
 @pytest.fixture()
