+++ vizro-core/tests/unit/vizro/integrations/kedro/test_kedro_data_manager.py
@@ -1,13 +1,20 @@
 """Unit tests for vizro.integrations.kedro."""
 import types
+from pathlib import Path
 
 import pytest
 import yaml
 
 kedro = pytest.importorskip("kedro")
+
 from kedro.io import DataCatalog  # noqa: E402
 
 from vizro.integrations.kedro import datasets_from_catalog  # noqa: E402
+
+
+@pytest.fixture
+def catalog_path():
+    return Path(__file__).parent / "fixtures/test_catalog.yaml"
 
 
 def test_datasets_from_catalog(catalog_path):
